pulled hasProperty() into own file, added documentation
