+++ lib/assertions.js
@@ -8,6 +8,7 @@ import isNotRequired from './assertions/is-not-required';
 import isVisible from './assertions/is-visible';
 import isNotVisible from './assertions/is-not-visible';
 import isDisabled from './assertions/is-disabled';
+import hasProperty from './assertions/has-property';
 
 import elementToString from './helpers/element-to-string';
 import collapseWhitespace from './helpers/collapse-whitespace';
@@ -193,6 +194,23 @@ export default class DOMAssertions {
    */
   isNotVisible(message) {
     isNotVisible.call(this, message);
+  }
+
+    /**
+   * Assert that the [HTMLElement][] or an [HTMLElement][] matching the
+   * `selector` contains the provided `property` and optionally checks
+   * if the property `value` matches the provided `value`.
+   *
+   * @param {string} property
+   * @param {object?} options
+   *
+   * @example
+   * assert.dom('.checkbox').hasProperty('checked');
+   * assert.dom('.checkbox').hasProperty('checked', { value: true });
+   * assert.dom('.checkbox').hasProperty('checked', { value: false, message: "The checkbox is not checked" });
+   */
+  hasProperty(property, options) {
+    hasProperty.call(this, property, options);
   }
 
   /**
@@ -659,38 +677,6 @@ export default class DOMAssertions {
 
   lacksValue(message) {
     this.hasNoValue(message);
-  }
-
-  hasProperty(property, options = {}) {
-    let { value, message } = options;
-    let element = this.findTargetElement();
-    if (!element) return;
-
-    const actualValue = element[property];
-    const propertyExists = typeof actualValue !== 'undefined';
-
-    if (typeof value === 'undefined') {
-      const result = propertyExists;
-      const expected = `Element ${this.targetDescription} has property "${property}"`;
-      const actual = result
-        ? expected
-        : `Element ${this.targetDescription} does not have property "${property}"`;
-      message = message || expected;
-
-      this.pushResult({ result, actual, expected, message });
-    } else {
-      const result = propertyExists && actualValue === value;
-      const expected = `Element ${this.targetDescription} has property "${property}" with value ${JSON.stringify(value)}`;
-      let actual;
-      if (!propertyExists) {
-        actual = `Element ${this.targetDescription} does not have property "${property}"`;
-      } else {
-        actual = `Element ${this.targetDescription} has property "${property}" with value ${JSON.stringify(actualValue)}`;
-      }
-      message = message || expected;
-
-      this.pushResult({ result, actual, expected, message });
-    }
   }
 
   /**
